Composed scoring/dedup helpers out of AbstractScoringTask into ScoringPipeline
* Moved RunCoelutionScoring, DeduplicateDoubleCounting, and DeduplicatePairs off the AbstractScoringTask base into a new composed ScoringPipeline collaborator in OspreySharp.Scoring, severing the PipelineContext dependency (logInfo delegate + IScoringDiagnostics, mirroring CoelutionScorer)
* AbstractScoringTask keeps thin protected forwarders so the six subclass call sites are byte-for-byte unchanged
* Added IScoringDiagnostics.DiagSearchEntryIds so the detected-entry count log no longer needs PipelineContext
* Removed the profiler-hook bracketing from RunCoelutionScoring (instrumentation only; placement to be revisited if needed)
* Pure code motion; Stellar regression byte-identical (modes 1 and 2), 390 tests, 0 warnings

See ai/todos/active/TODO-20260618_ospreysharp_debt_paydown_pr7.md
